+++ changelog.md
@@ -37,6 +37,16 @@
 
 
 
+* Added the owner of an item to the Loot History.
+* The owner has also been added to .csv and .tsv exports.
+* Note: Only items awarded after this release has their owners tracked.
+
+###### Dev
+* Changed parameters in ml_core `TrackAndLogLoot`.
+* Updated fields in `history_table` in said function, along with `RCMLLootHistorySend` message.
+
+
+
 ### v2.9.3
 ---
 * **Loot Status**